+++ bundles/aero.minova.rcp.rcp/src/aero/minova/rcp/rcp/parts/WFCDetailPart.java
@@ -4,6 +4,7 @@ package aero.minova.rcp.rcp.parts;
 import static aero.minova.rcp.rcp.fields.FieldUtil.TRANSLATE_PROPERTY;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -183,8 +184,6 @@ public class WFCDetailPart extends WFCFormPart {
 	private WFCDetailCASRequestsUtil casRequestsUtil;
 	private DirtyFlagUtil dirtyFlagUtil;
 
-	private IEclipseContext appContext;
-
 	private int detailWidth;
 
 	@Inject
@@ -208,10 +207,10 @@ public class WFCDetailPart extends WFCFormPart {
 
 	@PostConstruct
 	public void postConstruct(Composite parent, MWindow window, MApplication mApp) {
+
 		resManager = new LocalResourceManager(JFaceResources.getResources(), parent);
 		composite = parent;
 		formToolkit = new FormToolkit(parent.getDisplay());
-		appContext = mApp.getContext();
 		mApplication = mApp;
 		getForm();
 
@@ -230,6 +229,7 @@ public class WFCDetailPart extends WFCFormPart {
 
 		layoutForm(parent);
 
+		mDetail.setBooking("booking".equalsIgnoreCase(form.getDetail().getType()));
 		mDetail.setDetailAccessor(new DetailAccessor(mDetail));
 		ContextInjectionFactory.inject(mDetail.getDetailAccessor(), mPerspective.getContext()); // In Context, damit Injection verfügbar ist
 		mPerspective.getContext().set(MDetail.class, mDetail); // MDetail per Injection ermöglichen
@@ -1044,6 +1044,19 @@ public class WFCDetailPart extends WFCFormPart {
 
 			if (field.isVisible()) {
 				mSection.addMField(f);
+			}
+
+			// Wird ein primary-key geändert muss dies weitergegeben werden, damit Lookup-Masken funktionieren
+			if ("primary".equalsIgnoreCase(field.getKeyType())) {
+				f.addValueChangeListener(evt -> {
+					if (f.getValue() == null) {
+						return;
+					}
+					if (casRequestsUtil.getKeys() == null) {
+						casRequestsUtil.setKeys(new HashMap<>());
+					}
+					casRequestsUtil.getKeys().put(f.getName(), f.getValue());
+				});
 			}
 
 			return f;