+++ bundles/aero.minova.rcp.rcp/src/aero/minova/rcp/rcp/util/WFCDetailCASRequestsUtil.java
@@ -85,6 +85,9 @@ public class WFCDetailCASRequestsUtil {
 
 	private static final String ERROR = "Error";
 
+	private String updateRequest = Constants.UPDATE_REQUEST;
+	private String deleteRequest = Constants.DELETE_REQUEST;
+
 	@Inject
 	protected UISynchronize sync;
 
@@ -126,7 +129,7 @@ public class WFCDetailCASRequestsUtil {
 	@Inject
 	@Preference(nodePath = ApplicationPreferences.PREFERENCES_NODE, value = ApplicationPreferences.TIMEZONE)
 	public String timezone;
-	
+
 	@Inject
 	DirtyFlagUtil dirtyFlagUtil;
 
@@ -147,14 +150,18 @@ public class WFCDetailCASRequestsUtil {
 
 	private WFCDetailPart wfcDetailPart;
 
-	private Table currentKeyTable;
-
 	public void initializeCasRequestUtil(MDetail detail, MPerspective perspective, WFCDetailPart wfcDetailPart) {
 		this.mDetail = detail;
 		this.perspective = perspective;
 		this.wfcDetailPart = wfcDetailPart;
 		this.selectedOptionPages = new HashMap<>();
 		this.selectedGrids = new HashMap<>();
+
+		// Bei Buchungs-Toolbar wird Update zu Correct und Delete zu Cancel
+		if (mDetail.isBooking()) {
+			updateRequest = Constants.CORRECT_REQUEST;
+			deleteRequest = Constants.CANCEL_REQUEST;
+		}
 	}
 
 	/**
@@ -177,7 +184,6 @@ public class WFCDetailCASRequestsUtil {
 				broker.send(Constants.BROKER_CLEARSELECTION, perspective);
 				return;
 			}
-			currentKeyTable = table;
 
 			sendEventToHelper(ActionCode.BEFOREREAD);
 			updateGridLookupValues();
@@ -384,9 +390,9 @@ public class WFCDetailCASRequestsUtil {
 				MField f = mDetail.getField(name);
 				if (f != null) {
 					checkedFields.add(f);
-					if (f instanceof MQuantityField) {
+					if (f instanceof MQuantityField qf) {
 						QuantityValue value = new QuantityValue((Number) table.getValue(i, 0).getValue(),
-								table.getValue(((MQuantityField) f).getUnitFieldName(), 0).getStringValue());
+								table.getValue(qf.getUnitFieldName(), 0).getStringValue());
 						table.setValue(i, 0, value);
 					}
 					f.setValue(table.getValue(i, 0), false);
@@ -520,16 +526,13 @@ public class WFCDetailCASRequestsUtil {
 	}
 
 	private Table getInsertUpdateTable(Form buildForm) {
-		Table formTable;
-		if (getKeys() != null) {
-			formTable = dataFormService.getTableFromFormDetail(buildForm, Constants.UPDATE_REQUEST);
-		} else {
-			formTable = dataFormService.getTableFromFormDetail(buildForm, Constants.INSERT_REQUEST);
-			// Bei Insert wird OUTPUT gesetzt, damit die Keys des neu erstellten Eintrags zurückgegeben werden
+		Table formTable = dataFormService.getTableFromFormDetail(buildForm, getKeys() == null ? Constants.INSERT_REQUEST : updateRequest);
+
+		// Primary-Keys als Output definieren (für INSERT oder bei INSERT und CORRECT bei Buchung)
+		if (mDetail.isBooking() || getKeys() == null) {
 			for (aero.minova.rcp.model.Column c : formTable.getColumns()) {
 				String fieldname = (buildForm == form ? "" : buildForm.getDetail().getProcedureSuffix() + ".") + c.getName();
-				boolean isAField = mDetail.getField(fieldname) != null ? true : false;
-				if (isAField && mDetail.getField(fieldname).isPrimary()) {
+				if (mDetail.getField(fieldname) != null && mDetail.getField(fieldname).isPrimary()) {
 					c.setOutputType(OutputType.OUTPUT);
 				}
 			}
@@ -546,9 +549,9 @@ public class WFCDetailCASRequestsUtil {
 			sg.closeEditor();
 			sg.setPrimaryKeys(getKeys());
 
-			Table gridDeleteTable = TableBuilder.newTable(g.getProcedurePrefix() + Constants.DELETE_REQUEST + g.getProcedureSuffix()).create();
+			Table gridDeleteTable = TableBuilder.newTable(g.getProcedurePrefix() + deleteRequest + g.getProcedureSuffix()).create();
 			Table gridInsertTable = TableBuilder.newTable(g.getProcedurePrefix() + Constants.INSERT_REQUEST + g.getProcedureSuffix()).create();
-			Table gridUpdateTable = TableBuilder.newTable(g.getProcedurePrefix() + Constants.UPDATE_REQUEST + g.getProcedureSuffix()).create();
+			Table gridUpdateTable = TableBuilder.newTable(g.getProcedurePrefix() + updateRequest + g.getProcedureSuffix()).create();
 
 			for (aero.minova.rcp.model.Column gridColumn : g.getDataTable().getColumns()) {
 				aero.minova.rcp.model.Column c = new aero.minova.rcp.model.Column(gridColumn.getName(), gridColumn.getType());
@@ -568,13 +571,13 @@ public class WFCDetailCASRequestsUtil {
 			}
 
 			if (!gridDeleteTable.getRows().isEmpty()) {
-				procedureList.add(new TransactionEntry(Constants.GRID + "_" + g.getId() + "_" + Constants.DELETE_REQUEST, gridDeleteTable));
+				procedureList.add(new TransactionEntry(Constants.GRID + "_" + g.getId() + "_" + deleteRequest, gridDeleteTable));
 			}
 			if (!gridInsertTable.getRows().isEmpty()) {
 				procedureList.add(new TransactionEntry(Constants.GRID + "_" + g.getId() + "_" + Constants.INSERT_REQUEST, gridInsertTable));
 			}
 			if (!gridUpdateTable.getRows().isEmpty()) {
-				procedureList.add(new TransactionEntry(Constants.GRID + "_" + g.getId() + "_" + Constants.UPDATE_REQUEST, gridUpdateTable));
+				procedureList.add(new TransactionEntry(Constants.GRID + "_" + g.getId() + "_" + updateRequest, gridUpdateTable));
 			}
 		}
 	}
@@ -617,14 +620,9 @@ public class WFCDetailCASRequestsUtil {
 		}
 
 		SqlProcedureResult mainResult = resultList.get(0).getSQLProcedureResult();
-		if (insert) {
-			setKeysFromTable(mainResult.getOutputParameters());
-			broker.send(Constants.BROKER_SHOWNOTIFICATION, "msg.DataSaved");
-			handleUserAction(Constants.INSERT_REQUEST, mainResult.getOutputParameters());
-		} else {
-			broker.send(Constants.BROKER_SHOWNOTIFICATION, "msg.DataUpdated");
-			handleUserAction(Constants.UPDATE_REQUEST, currentKeyTable);
-		}
+		setKeysFromTable(mainResult.getOutputParameters());
+		broker.send(Constants.BROKER_SHOWNOTIFICATION, insert ? "msg.DataSaved" : "msg.DataUpdated");
+		handleUserAction(mainResult.getOutputParameters());
 
 		if (autoReloadIndex) {
 			ParameterizedCommand cmd = commandService.createCommand(Constants.AERO_MINOVA_RCP_RCP_COMMAND_LOADINDEX, null);
@@ -638,6 +636,11 @@ public class WFCDetailCASRequestsUtil {
 	 * @param t
 	 */
 	private void setKeysFromTable(Table t) {
+
+		if (t == null) {
+			return;
+		}
+
 		Map<String, Value> newKeys = new HashMap<>();
 		for (Field f : dataFormService.getAllPrimaryFieldsFromForm(form)) {
 			int index = t.getColumnIndex(f.getName());
@@ -651,12 +654,9 @@ public class WFCDetailCASRequestsUtil {
 	 *
 	 * @param updateRequest
 	 */
-	private void handleUserAction(String updateRequest, Table keyTable) {
-		Map<MPerspective, String> map = new HashMap<>();
-		map.put(perspective, updateRequest);
-
+	private void handleUserAction(Table keyTable) {
 		if (mDetail.isClearAfterSave()) {
-			clearFields(map);
+			clearFields("");
 		} else {
 			reloadFields(keyTable);
 		}
@@ -668,8 +668,7 @@ public class WFCDetailCASRequestsUtil {
 	@Inject
 	@Optional
 	public void reloadIndex(@UIEventTopic(Constants.BROKER_RELOADINDEX) String message) {
-		MPerspective activePerspective = model.getActivePerspective(partContext.get(MWindow.class));
-		if (activePerspective.equals(perspective)) {
+		if (isActivePerspective()) {
 			ParameterizedCommand cmd = commandService.createCommand(Constants.AERO_MINOVA_RCP_RCP_COMMAND_LOADINDEX, null);
 			handlerService.executeHandler(cmd);
 		}
@@ -703,7 +702,7 @@ public class WFCDetailCASRequestsUtil {
 				SectionGrid sg = ((GridAccessor) g.getGridAccessor()).getSectionGrid();
 				sg.closeEditor();
 
-				Table gridDeleteTable = TableBuilder.newTable(g.getProcedurePrefix() + Constants.DELETE_REQUEST + g.getProcedureSuffix()).create();
+				Table gridDeleteTable = TableBuilder.newTable(g.getProcedurePrefix() + deleteRequest + g.getProcedureSuffix()).create();
 				for (aero.minova.rcp.model.Column gridColumn : g.getDataTable().getColumns()) {
 					aero.minova.rcp.model.Column c = new aero.minova.rcp.model.Column(gridColumn.getName(), gridColumn.getType());
 					gridDeleteTable.addColumn(c);
@@ -731,26 +730,40 @@ public class WFCDetailCASRequestsUtil {
 	}
 
 	/**
-	 * Sucht die aktiven Controls aus der XMLDetailPart und baut anhand deren Werte eine Abfrage an den CAS zusammen
+	 * Baut die Tabelle für die Delete/Cancel-Anfrage
 	 */
 	private Table createDeleteTableFromForm(Form deleteForm) {
-		String tablename = deleteForm.getIndexView() != null ? "sp" : "op";
-		if ((!"sp".equals(deleteForm.getDetail().getProcedurePrefix()) && !"op".equals(deleteForm.getDetail().getProcedurePrefix()))) {
-			tablename = deleteForm.getDetail().getProcedurePrefix();
+		Table deleteTable;
+		// Bei Booking oder entsprechendem Flag in der Maske alle Felder übergeben
+		if (deleteForm.getDetail().isDeleteRequiresAllParams() || mDetail.isBooking()) { //
+			deleteTable = dataFormService.getTableFromFormDetail(deleteForm, deleteRequest);
+		} else { // Ansonsten nur primary-Keys
+			deleteTable = new Table();
+			deleteTable.setName(deleteForm.getDetail().getProcedurePrefix() + deleteRequest + deleteForm.getDetail().getProcedureSuffix());
+			for (Field f : dataFormService.getAllPrimaryFieldsFromForm(deleteForm)) {
+				String fieldName = (deleteForm == form ? "" : deleteForm.getDetail().getProcedureSuffix() + ".") + f.getName();
+				deleteTable.addColumn(new Column(f.getName(), mDetail.getField(fieldName).getDataType()));
+			}
 		}
-		tablename += "Delete";
-		tablename += deleteForm.getDetail().getProcedureSuffix();
-		TableBuilder tb = TableBuilder.newTable(tablename);
-		RowBuilder rb = RowBuilder.newRow();
-		for (Field f : dataFormService.getAllPrimaryFieldsFromForm(deleteForm)) {
-			String fieldName = (deleteForm == form ? "" : deleteForm.getDetail().getProcedureSuffix() + ".") + f.getName();
-			tb.withColumn(f.getName(), mDetail.getField(fieldName).getDataType());
-			rb.withValue(mDetail.getField(fieldName).getValue());
+
+		// Bei Booking Primary-Keys als Output definieren, damit Stornosatz geladen werden kann
+		if (mDetail.isBooking()) {
+			for (aero.minova.rcp.model.Column c : deleteTable.getColumns()) {
+				String fieldname = (deleteForm == form ? "" : deleteForm.getDetail().getProcedureSuffix() + ".") + c.getName();
+				if (mDetail.getField(fieldname) != null && mDetail.getField(fieldname).isPrimary()) {
+					c.setOutputType(OutputType.OUTPUT);
+				}
+			}
 		}
-		Table t = tb.create();
-		Row r = rb.create();
-		t.addRow(r);
-		return t;
+
+		Row r = new Row();
+		for (Column c : deleteTable.getColumns()) {
+			String fieldName = (deleteForm == form ? "" : deleteForm.getDetail().getProcedureSuffix() + ".") + c.getName();
+			r.addValue(mDetail.getField(fieldName).getValue());
+		}
+
+		deleteTable.addRow(r);
+		return deleteTable;
 	}
 
 	/**
@@ -772,9 +785,12 @@ public class WFCDetailCASRequestsUtil {
 			handlerService.executeHandler(cmd);
 		}
 		broker.send(Constants.BROKER_SHOWNOTIFICATION, "msg.DataDeleted");
-		Map<MPerspective, String> map = new HashMap<>();
-		map.put(perspective, Constants.DELETE_REQUEST);
-		clearFields(map);
+
+		if (mDetail.isBooking()) { // Bei Booking Stornosatz laden
+			reloadFields(resultList.get(0).getSQLProcedureResult().getOutputParameters());
+		} else {
+			clearFields("");
+		}
 		sendEventToHelper(ActionCode.AFTERDEL);
 		focusFirstEmptyField();
 	}
@@ -790,8 +806,7 @@ public class WFCDetailCASRequestsUtil {
 	@Optional
 	@Deprecated(since = "12.6.0", forRemoval = true)
 	public void callProcedureWithTable(@UIEventTopic(Constants.BROKER_PROCEDUREWITHTABLE) Table table) {
-		MPerspective activePerspective = model.getActivePerspective(partContext.get(MWindow.class));
-		if (activePerspective.equals(perspective) && table != null) {
+		if (isActivePerspective() && table != null) {
 
 			CompletableFuture<SqlProcedureResult> tableFuture = dataService.callProcedureAsync(table);
 
@@ -827,8 +842,8 @@ public class WFCDetailCASRequestsUtil {
 	 */
 	@Optional
 	@Inject
-	public void newFields(@UIEventTopic(Constants.BROKER_NEWENTRY) Map<MPerspective, String> map) {
-		if (map.keySet().iterator().next() != perspective) {
+	public void newFields(@UIEventTopic(Constants.BROKER_NEWENTRY) String s) {
+		if (!isActivePerspective()) {
 			return;
 		}
 
@@ -837,7 +852,7 @@ public class WFCDetailCASRequestsUtil {
 		}
 		updateGridLookupValues();
 		sendEventToHelper(ActionCode.BEFORENEW);
-		clearFields(map);
+		clearFields("");
 		// Helper-Klasse triggern, damit die Standard-Werte gesetzt werden können.
 		sendEventToHelper(ActionCode.AFTERNEW);
 		focusFirstEmptyField();
@@ -859,14 +874,26 @@ public class WFCDetailCASRequestsUtil {
 	}
 
 	/**
+	 * Direkt Methode clearFields(String s) benutzten. Diese Methode wird in einem Helper (für ScharrCPC) verwendet, vorerst behalten
+	 * 
+	 * @param map
+	 */
+	@Optional
+	@Inject
+	@Deprecated(since = "12.9.0", forRemoval = true)
+	public void clearFields(@UIEventTopic(Constants.BROKER_CLEARFIELDS) Map<MPerspective, String> map) {
+		clearFields("");
+	}
+
+	/**
 	 * Diese Methode bereiningt die Felder nach einer Erfolgreichen CAS-Anfrage
 	 *
 	 * @param origin
 	 */
 	@Optional
 	@Inject
-	public void clearFields(@UIEventTopic(Constants.BROKER_CLEARFIELDS) Map<MPerspective, String> map) {
-		if (map.keySet().iterator().next() != perspective) {
+	public void clearFields(@UIEventTopic(Constants.BROKER_CLEARFIELDS) String s) {
+		if (!isActivePerspective()) {
 			return;
 		}
 
@@ -877,19 +904,19 @@ public class WFCDetailCASRequestsUtil {
 		// Entfernen der Sub-Fields von den paramString Feldern
 		ArrayList<MField> paramfields = new ArrayList<>();
 		for (MField f : mDetail.getFields()) {
-			if (f instanceof MParamStringField) {
-				paramfields.addAll(((MParamStringField) f).getSubMFields());
+			if (f instanceof MParamStringField ps) {
+				paramfields.addAll(ps.getSubMFields());
 			}
 		}
 		mDetail.getFields().removeAll(paramfields);
 		// Felder auf Null setzen!
 		for (MField f : mDetail.getFields()) {
-			if (f instanceof MQuantityField) {
-				f.setUnitText(translationService.translate(((MQuantityField) f).getOriginalUnitText(), null));
+			if (f instanceof MQuantityField qf) {
+				f.setUnitText(translationService.translate(qf.getOriginalUnitText(), null));
 			}
 			f.setValue(null, false);
-			if (f instanceof MLookupField) {
-				((MLookupField) f).setOptions(null);
+			if (f instanceof MLookupField lf) {
+				lf.setOptions(null);
 			}
 		}
 
@@ -922,9 +949,10 @@ public class WFCDetailCASRequestsUtil {
 		for (MField field : mDetail.getFields()) {
 			if (field.getDefaultValueString() != null) {
 				String[] opAndFieldName = new String[2];
-				Value v = Value.getValueForStringFromDataType(field.getDefaultValueString(), field.getDataType(), field.getDateTimeType(), CustomLocale.getLocale(), timezone);
+				Value v = Value.getValueForStringFromDataType(field.getDefaultValueString(), field.getDataType(), field.getDateTimeType(),
+						CustomLocale.getLocale(), timezone);
 				field.setValue(v, false);
-				if(field.getName().contains(".")) {
+				if (field.getName().contains(".")) {
 					opAndFieldName = field.getName().split(".");
 				} else {
 					opAndFieldName[0] = null;
@@ -941,10 +969,15 @@ public class WFCDetailCASRequestsUtil {
 	@Inject
 	@Optional
 	public void reloadFields(@UIEventTopic(Constants.BROKER_RELOADFIELDS) Table keyTable) {
-		MPerspective activePerspective = model.getActivePerspective(partContext.get(MWindow.class));
-		if (activePerspective.equals(perspective)) {
+		if (isActivePerspective()) {
 			if (keyTable == null) {
-				keyTable = currentKeyTable;
+				keyTable = new Table();
+				Row r = new Row();
+				for (Entry<String, Value> e : keys.entrySet()) {
+					keyTable.addColumn(new Column(e.getKey(), e.getValue().getType()));
+					r.addValue(e.getValue());
+				}
+				keyTable.addRow(r);
 			}
 			readData(keyTable, false); // Nach Speichern soll discard-Nachricht nicht angezeigt werden
 		}
@@ -1054,9 +1087,9 @@ public class WFCDetailCASRequestsUtil {
 		// SubFelder von Param-String Feldern entfernen
 		List<MField> toRemove = new ArrayList<>();
 		for (MField mField : mDetail.getFields()) {
-			if (mField instanceof MParamStringField && mField.getMSection().equals(mSection)) {
-				toRemove.addAll(((MParamStringField) mField).getSubMFields());
-				((MParamStringField) mField).clearSubMFields();
+			if (mField instanceof MParamStringField psf && mField.getMSection().equals(mSection)) {
+				toRemove.addAll(psf.getSubMFields());
+				psf.clearSubMFields();
 			}
 		}
 		mSection.getMFields().removeAll(toRemove);
@@ -1072,11 +1105,10 @@ public class WFCDetailCASRequestsUtil {
 				visibleMFields.add(f);
 			}
 
-			if (f instanceof MParamStringField) {
+			if (f instanceof MParamStringField mParamString) {
 				String name = f.getName();
 				String suffix = name.contains("\\.") ? name.substring(name.lastIndexOf("\\."), name.length()) : "";
 
-				MParamStringField mParamString = (MParamStringField) f;
 				paramStringFields.add(mParamString);
 				for (Field subField : mParamString.getSubFields()) {
 					MField subMField = wfcDetailPart.createMField(subField, mSection, suffix);
@@ -1140,15 +1172,15 @@ public class WFCDetailCASRequestsUtil {
 			for (Object o : f.getDetail().getHeadAndPageAndGrid()) {
 				List<Object> fieldsOrGrids = new ArrayList<>();
 
-				if (o instanceof Head) {
-					fieldsOrGrids = ((Head) o).getFieldOrGrid();
-				} else if (o instanceof Page) {
-					fieldsOrGrids = ((Page) o).getFieldOrGrid();
+				if (o instanceof Head h) {
+					fieldsOrGrids = h.getFieldOrGrid();
+				} else if (o instanceof Page p) {
+					fieldsOrGrids = p.getFieldOrGrid();
 				}
 
 				for (Object fieldOrGrid : fieldsOrGrids) {
-					if (fieldOrGrid instanceof Field) {
-						subfields.add((Field) fieldOrGrid);
+					if (fieldOrGrid instanceof Field field) {
+						subfields.add(field);
 					}
 				}
 			}
@@ -1237,5 +1269,10 @@ public class WFCDetailCASRequestsUtil {
 	@Deprecated(since = "12.0.36", forRemoval = false)
 	public boolean checkDirty() {
 		return dirtyFlagUtil.checkDirty();
+	}
+
+	private boolean isActivePerspective() {
+		MPerspective activePerspective = model.getActivePerspective(partContext.get(MWindow.class));
+		return activePerspective.equals(perspective);
 	}
 }