+++ Changelog.md
@@ -15,6 +15,8 @@ Der entsprechende Abschnitt des Changelogs wird auch jeweils in die [Releasenote
 ## [Unreleased]
 
 ### Neu
+- Booking-Toolbar, aktiviert über Flag `type="booking"` in der Maske
+- Lookup-Masken nutzbar machen (nur Update Prozeduren, kein Insert)
 
 ### Änderung